fix(runtime): move synthetic baseline I/O runner out of @principles/core (PRI-206)
core now owns only pure contract/helpers:
- type exports (SyntheticBaselineSummary, stages, options)
- computeOverallStatus(), boundedEvidence(), safeStringify()
- truncateReason(), recommendNextIssue(), makeDeterministicDiagnosticianOutput()

I/O orchestration moved to pd-cli/src/services/synthetic-baseline-runner.ts:
- RuntimeStateManager, SQLite, PainSignalBridge, DiagnosticianRunner wiring
- queue and health read model checks
- temp workspace creation and resource cleanup

- reverted architecture-regression.test.ts trailing newline noise
- core tests now pure helper tests (23 tests, no fs/DB)
- CLI tests mock local runner service (9 tests)
