+++ packages/principles-core/src/runtime-v2/__tests__/architecture-regression.test.ts
@@ -3217,3 +3217,4 @@ describe('PRI-192 TraceRefinerAgent shadow contract boundary', () => {
     }
   });
 });
+
