+++ packages/principles-core/src/runtime-v2/index.ts
@@ -288,8 +288,8 @@ export type { CandidateAuditResult } from './candidate-audit.js';
 export { OperatorHealthReadModel } from './operator-health-read-model.js';
 export type { OperatorHealthSnapshot, OperatorHealthReadModelOptions, OverallHealthStatus } from './operator-health-read-model.js';
 
-// Synthetic baseline (PRI-206)
-export { runSyntheticBaseline, computeOverallStatus, boundedEvidence } from './synthetic-baseline.js';
+// Synthetic baseline (PRI-206) — pure contract/helpers only; I/O runner lives in pd-cli
+export { computeOverallStatus, boundedEvidence, safeStringify, truncateReason, makeDeterministicDiagnosticianOutput, recommendNextIssue } from './synthetic-baseline.js';
 export type { SyntheticBaselineSummary, SyntheticBaselineStage, SyntheticBaselineStageName, SyntheticBaselineFailStage, SyntheticBaselineOptions } from './synthetic-baseline.js';
 
 // Internalization contracts (PRI-42)
